Karma likes this form of expand better
